requirements.txt
aiogram>=3.22.0
asyncpg>=0.31.0
python-dotenv>=1.2.1
